Pull changes from VLM

+++ skeletons/asn_bit_data.c
@@ -86,7 +86,7 @@ asn_get_few_bits(asn_bit_data_t *pd, int nbits) {
 	else if(off <= 24)
 		accum = ((buf[0] << 16) + (buf[1] << 8) + buf[2]) >> (24 - off);
 	else if(off <= 31)
-		accum = ((buf[0] << 24) + (buf[1] << 16)
+		accum = (((uint32_t)buf[0] << 24) + (buf[1] << 16)
 			+ (buf[2] << 8) + (buf[3])) >> (32 - off);
 	else if(nbits <= 31) {
 		asn_bit_data_t tpd = *pd;
@@ -292,7 +292,7 @@ asn_put_aligned_flush(asn_bit_outp_t *po) {
         (po->buffer ? po->buffer - po->tmpspace : 0) + ((po->nboff + 7) >> 3);
 
     if(unused_bits) {
-        po->buffer[po->nboff >> 3] &= ~0 << unused_bits;
+        po->buffer[po->nboff >> 3] &= ~0u << unused_bits;
     }
 
     if(po->output(po->tmpspace, complete_bytes, po->op_key) < 0) {
